Moving patches/contenttypes.py to djangae.contrib.contenttypes
