+++ djangae/apps.py
@@ -6,9 +6,6 @@ class DjangaeConfig(AppConfig):
     verbose_name = _("Djangae")
 
     def ready(self):
-        from .patches.contenttypes import patch
-        patch(sender=self)
-
         from djangae.db.backends.appengine.caching import reset_context
         from django.core.signals import request_finished, request_started
