+++ djangae/contrib/contenttypes/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'djangae.contrib.contenttypes.apps.ContentTypesConfig'
